+++ src/lib/st-two-list-selection/list-scroll/list-scroll.component.ts
@@ -8,7 +8,8 @@
  *
  * SPDX-License-Identifier: Apache-2.0.
  */
-import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output } from '@angular/core';
+import { ChangeDetectorRef, ChangeDetectionStrategy, Component, ElementRef, EventEmitter, Input, Output, OnInit } from '@angular/core';
+import * as _ from 'lodash';
 
 import { StDropDownMenuItem } from '../../st-dropdown-menu/st-dropdown-menu.interface';
 import { StEgeo, StRequired } from '../../decorators/require-decorators';
@@ -20,22 +21,43 @@ import { StTwoListSelectionElement, StTwoListSelectExtraLabelAction } from '../s
    styleUrls: ['./list-scroll.component.scss'],
    changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class ListScrollComponent {
+export class ListScrollComponent implements OnInit {
 
    @Input() @StRequired() list: StTwoListSelectionElement[];
    @Input() editable: boolean = false;
+   @Input() fetchingDataText: string;
    @Input() hasAllList?: boolean = false;
+   @Input() isLoading?: boolean = false;
    @Input() itemAll?: StTwoListSelectionElement;
    @Input() @StRequired() qaTag: string;
    @Input() mode: 'compact' | 'normal' = 'normal';
 
+   @Output() scrollBottom: EventEmitter<any> = new EventEmitter<any>();
    @Output() selectItem: EventEmitter<StTwoListSelectionElement> = new EventEmitter<StTwoListSelectionElement>();
    @Output() selectExtraLabel: EventEmitter<StTwoListSelectExtraLabelAction> = new EventEmitter<StTwoListSelectExtraLabelAction>();
    @Output() search: EventEmitter<string> = new EventEmitter<string>();
 
-
-   scrollItems: StTwoListSelectionElement[] = [];
+   atBottom: Boolean = false;
+   delay: any = (() => {
+      let timer = 0;
+      return (callback: any, ms: any): void => {
+         clearTimeout(timer);
+         timer = setTimeout(callback, ms);
+      };
+   })();
    firstEl: number = 0;
+   scrollItems: StTwoListSelectionElement[] = [];
+   showLoading: Boolean = false;
+
+   constructor(
+      private cd: ChangeDetectorRef,
+      private elemRef: ElementRef) {
+   }
+
+   ngOnInit(): void {
+      this.showLoading = this.isLoading;
+      this.cd.markForCheck();
+   }
 
    get listQaTag(): string {
       return this.qaTag + '-scroll-list';
@@ -45,4 +67,17 @@ export class ListScrollComponent {
       return this.qaTag + '-check-all-scroll-list';
    }
 
+   onScroll(): void {
+      let element = this.elemRef.nativeElement.querySelector('#virtualScroll');
+      this.showLoading = this.isLoading;
+      this.atBottom = element.scrollHeight - element.scrollTop === element.clientHeight;
+
+      const debounce = _.throttle(() => {
+         if (this.atBottom) {
+            this.scrollBottom.emit();
+            this.showLoading = false;
+         }
+      });
+      this.delay(debounce, 1000);
+   }
 }